requirements.txt
# asgiref==3.8.1
Django==5.1.2
django-cors-headers==4.5.0
mysql-connector-python==9.0.0
# psycopg2-binary==2.9.10
# sqlparse==0.5.1
# mysqlclient==2.2.5
# mysql-connector-python==9.0.0
gunicorn==22.0.0
uvicorn==0.30.1
djangorestframework==3.14.0
python-decouple==3.6
whitenoise==6.2.0
# mysqlclient==2.2.5
django-jazzmin==2.5.0
Pillow==9.2.0